Implement repeated simulations with many selfish
Abstract single simulation in pvs.py, add utils.py with
plot_and_save(), add more figures

By changing the parameters in pvs.py and invoking `python pvs.py`,
a single voting system simulation with the desired parameters is run
(just like with main.py prior to this commit). The functionality of
main.py is now to run many simulations (with the parameters as set
in main.py) for a variable number of Sybil accounts.